+++ src/App.jsx
@@ -6,7 +6,7 @@ import Admin from "./pages/Admin.jsx";
 import Login from './pages/Login.jsx';
 import Register from './pages/Register.jsx';
 import ContextProvider from "./utils/global.context.jsx";
-import AdminAsidePrueba from "./pages/AdminAsidePrueba.jsx";
+
 
 function App() {
   return (
@@ -19,7 +19,6 @@ function App() {
               <Route path="/administracion" element={<Admin />} />
               <Route path="/login" element={<Login />} />
               <Route path="/register" element={<Register />} />
-              <Route path="/prueba" element={<AdminAsidePrueba />} />
             </Route>
           </Routes>
         </Router>
