+++ src/pages/Admin.jsx
@@ -1,90 +1,339 @@
 import { useContextGlobal } from "../utils/global.context";
 import SubHeader from "../components/SubHeader";
 import ProductTable from "../components/admin/ProductTable";
-import Form from "../components/admin/Form";
+import UserTable from "../components/admin/UserTable";
+import CategoryTable from "../components/admin/CategoryTable";
 import Message from "../components/admin/Message";
 import IsMobile from "../components/admin/IsMobile";
 import { useState } from "react";
+import Sidebar from "../components/admin/Sidebar";
+import Form from "../components/admin/Form";
+import { FaTimes } from "react-icons/fa";
+import { idCreator } from "../utils/formatFunctions";
 
 const Admin = () => {
-    const { isMobile } = useContextGlobal();
-    const [isCreatingProduct, setIsCreatingProduct] = useState(false);
-    const [successMessage, setSuccessMessage] = useState("");
-    const [errorMessage, setErrorMessage] = useState("");
+	const { isMobile, state, dispatch } = useContextGlobal();
+	const [isCreatingItem, setIsCreatingItem] = useState(null); // Maneja qué formulario se está mostrando (producto, usuario o categoría)
+	const [successMessage, setSuccessMessage] = useState("");
+	const [errorMessage, setErrorMessage] = useState("");
+	const [newUser, setNewUser] = useState({
+		id: "",
+		nombre: "",
+		apellido: "",
+		email: "",
+		rol: "admin",
+	});
+	const handleAddItem = (itemType) => {
+		setIsCreatingItem(itemType); // Establece el tipo de ítem que se va a crear
+	};
 
-    const handleAddProduct = () => {
-        setIsCreatingProduct(true);
-    };
+	const handleListItems = () => {
+		setIsCreatingItem(null); // Vuelve a la vista de lista
+	};
 
-    const handleListProducts = () => {
-        setIsCreatingProduct(false);
-    };
+	const activeSection = state.activeSection;
 
-    const buttons = {
-        1: [
-            {
-                text: "Agregar producto",
-                textColor: "primary",
-                bgColor: "transparent",
-                action: handleAddProduct,
-            },
-        ],
-        2: [
-            {
-                text: "Lista de productos",
-                textColor: "primary",
-                bgColor: "transparent",
-                action: handleListProducts,
-            },
-        ],
-    };
+	const buttons = {
+		obras: [
+			{
+				text: "Agregar producto",
+				textColor: "primary",
+				bgColor: "transparent",
+				action: () => handleAddItem("producto"),
+			},
+			{
+				text: "Lista de productos",
+				textColor: "primary",
+				bgColor: "transparent",
+				action: handleListItems,
+			},
+		],
+		usuarios: [
+			{
+				text: "Agregar usuario",
+				textColor: "primary",
+				bgColor: "transparent",
+				action: () => handleAddItem("usuario"),
+			},
+			{
+				text: "Lista de usuarios",
+				textColor: "primary",
+				bgColor: "transparent",
+				action: handleListItems,
+			},
+		],
+		categorias: [
+			{
+				text: "Agregar categoría",
+				textColor: "primary",
+				bgColor: "transparent",
+				action: () => handleAddItem("categoria"),
+			},
+			{
+				text: "Lista de categorías",
+				textColor: "primary",
+				bgColor: "transparent",
+				action: handleListItems,
+			},
+		],
+	};
 
-    return (
-        <>
-            {isMobile ? (
-                <IsMobile />
-            ) : (
-                <div className="min-h-screen pt-8 bg-black">
-                    <SubHeader title={"Panel de Administración"} buttons={buttons} />
-                    {isCreatingProduct ? (
-                        <section className="py-24 bg-black h-full mx-24 mt-20">
-                            <Form
-                                edit={false}
-                                onClose={handleListProducts}
-                                setSuccessMessage={setSuccessMessage}
-                                setErrorMessage={setErrorMessage}
-                            />
-                        </section>
-                    ) : (
-                        <ProductTable
-                            setSuccessMessage={setSuccessMessage}
-                            setErrorMessage={setErrorMessage}
-                        />
-                    )}
+	const buttonsToDisplay = { [activeSection]: buttons[activeSection] || [] };
 
-                    {/* Success or error message */}
-                    {successMessage && (
-                        <div className="fixed bottom-4 right-4 z-50 mb-4">
-                            <Message
-                                type="success"
-                                text={successMessage}
-                                onClose={() => setSuccessMessage("")}
-                            />
-                        </div>
-                    )}
-                    {errorMessage && (
-                        <div className="fixed bottom-4 right-4 z-50 mb-4">
-                            <Message
-                                type="danger"
-                                text={errorMessage}
-                                onClose={() => setErrorMessage("")}
-                            />
-                        </div>
-                    )}
-                </div>
-            )}
-        </>
-    );
+	const handleInputChange = (e) => {
+		const { name, value } = e.target;
+		setNewUser({
+			...newUser,
+			[name]: value,
+		});
+	};
+
+	const handleSubmitUser = (e) => {
+		e.preventDefault(); // Previene el comportamiento predeterminado de envío del formulario
+
+		if (!newUser.nombre || !newUser.apellido || !newUser.email) {
+			setErrorMessage("Por favor, complete todos los campos.");
+			return;
+		}
+
+		// Asignar el ID
+		const newUserWithId = {
+			...newUser,
+			id: idCreator(state.users),
+		};
+
+		// Agregar el nuevo usuario
+        console.log("admin: ",newUserWithId)
+		dispatch({ type: "ADD_USER", payload: newUserWithId });
+		setSuccessMessage("Usuario creado con éxito");
+
+		// Limpiar los campos después de la creación
+		setNewUser({ nombre: "", apellido: "", email: "", rol: "user" });
+        
+		handleListItems();
+	};
+
+	return (
+		<>
+			{isMobile ? (
+				<IsMobile />
+			) : (
+				<div className="min-h-screen pt-8 bg-black">
+					<SubHeader
+						title={"Panel de Administración"}
+						buttons={buttonsToDisplay}
+					/>
+					{isCreatingItem ? (
+						<section className="flex w-screen h-screen-28">
+							<Sidebar />
+							<div className="flex flex-col items-center grow max-h-screen pt-32 relative">
+								{isCreatingItem === "producto" && (
+									<Form
+										edit={false}
+										onClose={handleListItems}
+										setSuccessMessage={setSuccessMessage}
+										setErrorMessage={setErrorMessage}
+									/>
+								)}
+								{isCreatingItem === "usuario" && (
+									<div className="w-[75vw] h-[70vh] overflow-y-scroll bg-white p-6 rounded-lg shadow-md relative">
+										<button
+											onClick={handleListItems}
+											className="absolute top-2 right-2 bg-red-600 text-white rounded-full w-8 h-8 flex items-center justify-center"
+											aria-label="Cerrar"
+										>
+											<FaTimes />
+										</button>
+										<h2 className="text-xl font-semibold mb-4">
+											Crear nuevo usuario
+										</h2>
+										<form onSubmit={handleSubmitUser}>
+											<div className="flex space-x-4 mb-4">
+												<div className="flex-1">
+													<label
+														className="block text-sm font-semibold mb-2"
+														htmlFor="nombre"
+													>
+														Nombre
+													</label>
+													<input
+														type="text"
+														id="nombre"
+														name="nombre"
+														value={newUser.nombre}
+														onChange={
+															handleInputChange
+														}
+														className="w-full p-2 border border-gray-300 rounded"
+														required
+													/>
+												</div>
+												<div className="flex-1">
+													<label
+														className="block text-sm font-semibold mb-2"
+														htmlFor="apellido"
+													>
+														Apellido
+													</label>
+													<input
+														type="text"
+														id="apellido"
+														name="apellido"
+														value={newUser.apellido}
+														onChange={
+															handleInputChange
+														}
+														className="w-full p-2 border border-gray-300 rounded"
+														required
+													/>
+												</div>
+											</div>
+											<div className="mb-4">
+												<label
+													className="block text-sm font-semibold mb-2"
+													htmlFor="email"
+												>
+													Correo electrónico
+												</label>
+												<input
+													type="email"
+													id="email"
+													name="email"
+													value={newUser.email}
+													onChange={handleInputChange}
+													className="w-full p-2 border border-gray-300 rounded"
+													required
+												/>
+											</div>
+											<div className="mb-4">
+												<label
+													className="block text-sm font-semibold mb-2"
+													htmlFor="rol"
+												>
+													Rol
+												</label>
+												<select
+													id="rol"
+													name="rol"
+													value={newUser.rol}
+													onChange={handleInputChange}
+													className="w-full p-2 border border-gray-300 rounded"
+													required
+												>
+													<option value="admin">
+														Administrador
+													</option>
+													<option value="colab">
+														Colaborador
+													</option>
+													<option value="user">
+														Usuario
+													</option>
+												</select>
+											</div>
+											<div className="flex justify-between">
+												<button
+													type="button"
+													className="bg-gray-500 text-white py-2 px-4 rounded"
+													onClick={handleListItems}
+												>
+													Cancelar
+												</button>
+												<button
+													type="submit"
+													className="bg-blue-600 text-white py-2 px-4 rounded"
+												>
+													Crear Usuario
+												</button>
+											</div>
+										</form>
+									</div>
+								)}
+								{isCreatingItem === "categoria" && (
+									<div className="w-[75vw] h-[70vh] overflow-y-scroll bg-white p-6 rounded-lg shadow-md relative">
+										<button
+											onClick={handleListItems}
+											className="absolute top-2 right-2 bg-red-600 text-white rounded-full w-8 h-8 flex items-center justify-center"
+											aria-label="Cerrar"
+										>
+											<FaTimes />
+										</button>
+										<h2 className="text-xl font-semibold mb-4">
+											Crear nueva categoría
+										</h2>
+										<form>
+											<div className="mb-4">
+												<label
+													className="block text-sm font-semibold mb-2"
+													htmlFor="nombre"
+												>
+													Nombre de la categoría
+												</label>
+												<input
+													type="text"
+													id="nombre"
+													className="w-full p-2 border border-gray-300 rounded"
+												/>
+											</div>
+											<div className="mb-4">
+												<label
+													className="block text-sm font-semibold mb-2"
+													htmlFor="descripcion"
+												>
+													Descripción
+												</label>
+												<textarea
+													id="descripcion"
+													className="w-full p-2 border border-gray-300 rounded"
+												/>
+											</div>
+											<div className="mb-4">
+												<label
+													className="block text-sm font-semibold mb-2"
+													htmlFor="imagen"
+												>
+													Imagen (URL)
+												</label>
+												<input
+													type="text"
+													id="imagen"
+													className="w-full p-2 border border-gray-300 rounded"
+												/>
+											</div>
+											<div className="flex justify-between">
+												<button
+													type="button"
+													className="bg-gray-500 text-white py-2 px-4 rounded"
+													onClick={handleListItems}
+												>
+													Cancelar
+												</button>
+												<button
+													type="submit"
+													className="bg-blue-600 text-white py-2 px-4 rounded"
+												>
+													Crear Categoría
+												</button>
+											</div>
+										</form>
+									</div>
+								)}
+							</div>
+						</section>
+					) : (
+						<section className="flex w-screen h-screen-28">
+							<Sidebar />
+							{activeSection === "obras" && <ProductTable />}
+							{activeSection === "usuarios" && <UserTable />}
+							{activeSection === "categorias" && (
+								<CategoryTable />
+							)}
+						</section>
+					)}
+				</div>
+			)}
+		</>
+	);
 };
 
 export default Admin;
