+++ src/reducers/reducer.js
@@ -59,20 +59,23 @@ export const reducer = (state, action) => {
 			saveToLocalStorage("categories", filteredCategories);
 			return { ...state, categories: filteredCategories };
 
-			case "ADD_USER":
-				return { ...state, users: [...state.users, action.payload] };
-			  case "UPDATE_USER":
-				return {
-				  ...state,
-				  users: state.users.map((user) =>
+		case "ADD_USER":
+			console.log("dispatch: ",action.payload)
+			return { ...state, users: [...state.users, action.payload] };
+
+		case "UPDATE_USER":
+			return {
+				...state,
+				users: state.users.map((user) =>
 					user.id === action.payload.id ? action.payload : user
-				  ),
-				};
-			  case "DELETE_USER":
-				return {
-				  ...state,
-				  users: state.users.filter((user) => user.id !== action.payload.id),
-				};
+				),
+			};
+		case "DELETE_USER":
+			const filteredDataUser = state.users.filter(
+				(item) => item.id !== action.payload.id
+			);
+			saveToLocalStorage("users", filteredDataUser);
+			return { ...state, users: filteredDataUser };
 
 		case "CHANGE_THEME":
 			return { ...state, theme: action.payload };
@@ -80,44 +83,43 @@ export const reducer = (state, action) => {
 		case "SET_ACTIVE_SECTION":
 			console.log("Cambiando sección a reducer:", action.payload); // Verifica que el payload sea el correcto
 			return { ...state, activeSection: action.payload };
-			case "ADD_IMAGE":
-				const newImages = [...state.images, action.payload]; // action.payload puede ser una URL o base64
-				saveToLocalStorage("images", newImages); // Guarda las imágenes en localStorage
-				return { ...state, images: newImages };
-			
-			case "UPDATE_IMAGE":
-				const updatedImages = state.images.map((image) =>
-					image.id === action.payload.id
-						? { ...image, ...action.payload }
-						: image
-				);
-				saveToLocalStorage("images", updatedImages);
-				return { ...state, images: updatedImages };
-			
-			case "DELETE_IMAGE":
-				const filteredImages = state.images.filter(
-					(image) => image.id !== action.payload.id
-				);
-				saveToLocalStorage("images", filteredImages);
-				return { ...state, images: filteredImages };
-			
-			case "ADD_IMAGE_TO_ART":
-				return {
-					...state,
-					data: state.data.map((obra) =>
-						obra.id === action.payload.artId
-							? {
+		case "ADD_IMAGE":
+			const newImages = [...state.images, action.payload]; // action.payload puede ser una URL o base64
+			saveToLocalStorage("images", newImages); // Guarda las imágenes en localStorage
+			return { ...state, images: newImages };
+
+		case "UPDATE_IMAGE":
+			const updatedImages = state.images.map((image) =>
+				image.id === action.payload.id
+					? { ...image, ...action.payload }
+					: image
+			);
+			saveToLocalStorage("images", updatedImages);
+			return { ...state, images: updatedImages };
+
+		case "DELETE_IMAGE":
+			const filteredImages = state.images.filter(
+				(image) => image.id !== action.payload.id
+			);
+			saveToLocalStorage("images", filteredImages);
+			return { ...state, images: filteredImages };
+
+		case "ADD_IMAGE_TO_ART":
+			return {
+				...state,
+				data: state.data.map((obra) =>
+					obra.id === action.payload.artId
+						? {
 								...obra,
 								imagenesAdicionales: [
 									...obra.imagenesAdicionales,
 									action.payload.imgUrl,
 								],
-							}
-							: obra
-					),
-				};
+						  }
+						: obra
+				),
+			};
 
-			
 		default:
 			throw new Error("Acción no existente");
 	}
